+++ server/tests/unit/services/checkCine/cinemaFunctions.test.js
@@ -1,16 +1,16 @@
-require("../../../src/utils/errorLogger");
+require("../../../../src/utils/errorLogger");
 const {
     cinemaC,
     cinemaL,
     cinemaM,
-} = require("../../../src/services/v1/checkCine/cinema");
-const cinemaLModel = require("../../../src/models/cinema/cinemaL");
+} = require("../../../../src/services/v1/checkCine/cinema");
+const cinemaLModel = require("../../../../src/models/cinema/cinemaL");
 
 const axios = require("axios");
 const FormData = require("form-data");
 jest.mock("axios");
 jest.mock("form-data");
-jest.mock("../../../src/models/cinema/cinemaL");
+jest.mock("../../../../src/models/cinema/cinemaL");
 
 global.bodyGenerator = jest.fn();
